Widen write and measure bit counters to 64-bit (#68)
Johnny Grok. The write and measure streams count bits in serialize_int64_t, the accessors return it, and every byte operand is cast before it is multiplied, so a 256 MiB buffer no longer overflows the counter; roundtrip.c pins both the write and measure sides. Read cold today; the reader's fixes are on the branch; CI green. Closes #67.

+++ test/conformance.c
@@ -1587,8 +1587,8 @@ static void run_measure( const vector_t * v, step_t * steps, int step_count )
     if ( (serialize_int64_t) serialize_measure_bits_processed( &m ) < v->measure_at_least )
     {
         char message[MAX_LINE];
-        sprintf( message, "measured %d bits, the corpus requires at least %ld",
-                 serialize_measure_bits_processed( &m ), (long) v->measure_at_least );
+        sprintf( message, "measured %lld bits, the corpus requires at least %ld",
+                 (long long) serialize_measure_bits_processed( &m ), (long) v->measure_at_least );
         fail( v, message );
     }
 }
